all code workinggit add .! All code cleaned upgit add .! Time to finish style
