+++ src/Components/layout/buttons/leftSidebutton.js
@@ -1,7 +1,7 @@
-import React, { Component } from 'react';
-import { createMuiTheme, withStyles, makeStyles, ThemeProvider } from '@material-ui/core/styles';
+import React from 'react';
+import {  withStyles, makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import API from "../../../utils/API";
+// import API from "../../../utils/API";
 
 const BootstrapButton = withStyles({
   root: {
