+++ src/Components/layout/cards/IntroCard.js
@@ -4,11 +4,9 @@ import CardContent from "@material-ui/core/CardContent";
 import CardMedia from "@material-ui/core/CardMedia";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
 import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
 import Box from "@material-ui/core/Box";
-import { shadows } from '@material-ui/system';
+
 
 const useStyles = makeStyles({
   root: {
