+++ src/Components/layout/navbar/NavBar.js
@@ -4,55 +4,40 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import LoginButton from '../buttons/LoginButton';
 import CustomizedInputBase from '../search/search';
-import Logo from '../../../assets/Logos/Logo'
+import Logo from '../../../assets/Logos/Logo';
 import { createMuiTheme } from '@material-ui/core/styles';
 import { ThemeProvider } from '@material-ui/styles';
-import { makeStyles } from '@material-ui/core/styles';
+import AccountBtn from '../buttons/accountBtn';
 
 
-// const theme = createMuiTheme({
-//   grow: {
-//     flexGrow: 1,
-//   },
-//   overrides: {
-//     MuiAppBar: {
-//       colorPrimary: {
-//         backgroundColor: '#1C1E1E',
-//         alignItems: 'space-between',
-//       },
-//     },
-
+const theme = createMuiTheme({
+  overrides: {
+    MuiAppBar: {
+      colorPrimary: {
+        backgroundColor: '#1C1E1E',
+      },
+    },
     
-//   },
-// });
+  },
+});
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-  },
-}));
-export default function ButtonAppBar() {
-  const classes = useStyles();
+export default function ButtonAppBar(props) {
+
 
   return (
-    // <ThemeProvider theme={theme}>
-    <div className={classes.root}>
-      <AppBar position="static" disableGutters>
+    <ThemeProvider theme={theme}>
+      <div>
+      <AppBar  postition= "static" disableGutters>
         <Toolbar>
           <Typography variant="h6">
-            <Logo edge="start" className={classes.menuButton}/>
+            <Logo edge= "start"/>
           </Typography>
-          <CustomizedInputBase maxWidth="300px" alignItems="right"/>
+          <CustomizedInputBase getQueryTopic={props.getQueryTopic} maxWidth="300px"/>
           <LoginButton />
+          <AccountBtn/>
         </Toolbar>
       </AppBar>
-    </div>
-    // </ThemeProvider>
+      </div>
+    </ThemeProvider>
   );
 }
